[WIP] WAN protocol experiment:
- use an operation flag as indication AuthorizationOp sends
additional field supportedProtocols
- when receiving a new style AuthorizationOp, respond with the
first protocol in the supportedProtocols array, otherwise respond with
a boolean (legacy protocol)
- on the receiving end (in WanConnectionManager), detect whether the
response is from a legacy or new Hazelcast member; if a protocol is
selected, send a dummy op; if legacy protocol is selected, just return
the connection.

This has been tested (manually) and works as expected with WAN targets
on a 3.8.3 member and a 3.11-SNAPSHOT (with this commit) member,
following the appropriate code path each time.