[HZ-264] Differential partition sync
With differential partition sync it is possible
that some existing keys are overwritten with
same or updated value, since differential sync
does not perform exact comparison of what changed
or not. In this case, we need to update the existing
record instead of always creating a new one, otherwise
hot restart loses track of key handles. To this end
putOrUpdateReplicatedRecord method is introduced
in RecordStore interface.